Parse split-dns information from xml
Add a new split_dns structure to store split_dns info for
the tunnel. First step towards getting for support for
split dns in the long run. The device available to test
only had one domain and one server configured, so not sure
how multiple will look like, so for now the full domain
string is just saved as string. dnsserver is populated
in a loop, max of 2 supported for now.