Add workaround for split-dns
Split-dns seems to be quite a complicated thing to implement
correctly, and probably depends on something like systemd-resolvd
or NetworkManager support to be added, as discussed here:
    https://github.com/adrienverge/openfortivpn/issues/561

In the mean time DNS is not working at all if no global dns
is set. This patch implements a slightly ugly workaround by using
the split-dns info as the default dns when no default dns was configured
in the xml response. I expect that the two settings are mutually exclusive in
any case, but I don't understand DNS well enough to be sure.